feat: 라우터 기능 추가 및 Home 페이지 분리
