+++ README.md
@@ -34,8 +34,8 @@
 
 ### 5. ✅ 카드 등록 완료 및 네비게이션
 
-- [ ] 확인 버튼을 클릭하면 사용자는 카드 등록 완료 페이지로 이동한다.
-- [ ] 카드 등록 완료 페이지에서는 카드 등록이 성공적으로 완료되었다는 메시지와 함께, 다시 카드 등록 페이지로 돌아갈 수 있는 확인 버튼을 제공한다.
+- [x] 확인 버튼을 클릭하면 사용자는 카드 등록 완료 페이지로 이동한다.
+- [x] 카드 등록 완료 페이지에서는 카드 등록이 성공적으로 완료되었다는 메시지와 함께, 다시 카드 등록 페이지로 돌아갈 수 있는 확인 버튼을 제공한다.
 
 <br>
 
@@ -72,9 +72,9 @@
 ## [Step2] ✅ 프로그래밍 요구사항
 
 - Routing
-  - [ ] react-router를 이용하여 페이지를 관리한다.
+  - [x] react-router를 이용하여 페이지를 관리한다.
 - Custom Hooks
-  - [ ] custom hooks를 활용하여 form 관리에 필요한 로직을 UI와 분리하여 관리한다. 이를 통해 코드의 가독성과 재사용성을 높인다.
+  - [x] custom hooks를 활용하여 form 관리에 필요한 로직을 UI와 분리하여 관리한다. 이를 통해 코드의 가독성과 재사용성을 높인다.
 - Test
   - [x] 각 스토리에 명확한 이름을 지정하고, 스토리 이름을 통해 컴포넌트의 사용 사례를 쉽게 이해할 수 있도록 한다.
